Add utility to run API and frontend
